Rewrite FdEntry reusing as much libstd as possible
