+++ src/sys/mod.rs
@@ -1,13 +1,13 @@
-use cfg_if::cfg_if;
+// use cfg_if::cfg_if;
 
-cfg_if! {
-    if #[cfg(unix)] {
-        mod unix;
-        pub use self::unix::*;
-    } else if #[cfg(windows)] {
-        mod windows;
-        pub use self::windows::*;
-    } else {
-        compile_error!("wasi-common doesn't compile for this platform yet");
-    }
-}
+// cfg_if! {
+//     if #[cfg(unix)] {
+mod unix;
+pub use self::unix::*;
+// } else if #[cfg(windows)] {
+//     mod windows;
+//     pub use self::windows::*;
+// } else {
+//     compile_error!("wasi-common doesn't compile for this platform yet");
+// }
+// }
