+++ src/sys/unix/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod fdentry;
+pub(crate) mod fdentry_impl;
 pub(crate) mod host_impl;
 pub(crate) mod hostcalls_impl;
